rename auth operations
