+++ src/pages/UserPage.js
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { fetchPets } from 'redux/pets/operations';
-import { hideModalSuccessRegister } from 'redux/auth/operetions';
+import { hideModalSuccessRegister } from 'redux/auth/operations';
 import { useSelector } from 'react-redux';
 
 import UserData from 'components/UserData';
